% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensors.R
\name{get_sensors}
\alias{get_sensors}
\title{Retrieve list of available sensors for given station from GIOS API}
\usage{
get_sensors(stationId)
}
\arguments{
\item{stationId}{id of a station}
}
\value{
data.frame all sensors available for a given station
}
\description{
Retrieve list of available sensors for given station from GIOS API
}
\examples{
\dontrun{
stations <- get_stations()
subset(stations, city.name == "Wrocław")
get_sensors(129)
}

}
