% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosAPI-package.R
\docType{package}
\name{giosAPI}
\alias{giosAPI}
\alias{giosAPI-package}
\title{Interface to the GIOS API}
\description{
A set of tools that provide access to the GIOS air quality data API.
For more information please visit \url{http://powietrze.gios.gov.pl/pjp/content/api}.
}
\author{
Piotr Sobczyk (pj.sobczyk@gmail.com)
}
