% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurment_data.R
\name{get_measurment_data}
\alias{get_measurment_data}
\title{Retrieve measurment data for a given sensor from GIOS API}
\usage{
get_measurment_data(sensorId)
}
\arguments{
\item{sensorId}{id of a sensor}
}
\value{
data.frame last 24h data
}
\description{
Retrieve measurment data for a given sensor from GIOS API
}
